Refatoração, segurança e novas páginas Blazor RifaTech
Refatoração de arquitetura, melhorias de segurança (CORS, JWT, MercadoPago, rate limiting), modelagem de dados com decimal e validação FluentValidation, unificação de serviços, seed programático, testes unitários, health checks e Swagger aprimorado.
Endpoints minimal API validados, agrupados e com logs estruturados.
Frontend Blazor: novas páginas (admin, tickets, rifas, ajuda), tratamento de imagens, página "Not Found" e melhorias de UI.
Infraestrutura Dockerfile e docker-compose para API/UI/MySQL.
Documentação de avaliação arquitetural.
Middleware global de exceções, endpoint de config WASM e atualização de pacotes.
Projeto pronto para produção, seguro, escalável e com interface completa.

+++ RifaTech.API/Repositories/PaymentService.cs
@@ -216,7 +216,7 @@ namespace RifaTech.API.Repositories
                 {
                     ClienteId = cliente.Id,
                     Method = "PIX",
-                    Amount = (float)valorTotal,
+                    Amount = valorTotal,
                     IsConfirmed = false,
                     QrCodeBase64 = mercadoPagoPayment.PointOfInteraction?.TransactionData?.QrCodeBase64 ?? "",
                     QrCode = mercadoPagoPayment.PointOfInteraction?.TransactionData?.QrCode ?? "",
